pyproject.toml
[tool.poetry]
name = "cars-control"
version = "0.1.0"
description = ""
authors = ["Grigoriy1209 <grigoriyv1209@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.5"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
requests = "^2.32.3"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
drf-yasg = "^1.21.8"
drf-nested-routers = "^0.94.1"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.2.1"
daphne = "^4.1.2"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
